enterprises/pyproject.toml
[tool.poetry]
name = "enterprises"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework-simplejwt = "^5.2.1"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
pytest-django = "^4.5.2"
pytest = "^7.1.3"
requests = "^2.28.1"
psycopg2 = "^2.9.4"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
